refactor(shared): extract doubao provider metadata
